[test_techsupport]: Add show techsupport UTs to KVM test. (#4912)
### Description of PR
    Add show techsupport UTs  to KVM test.

### Type of change
- [ ] Bug fix
- [ ] Testbed and Framework(new/improvement)
- [x] Test case(new/improvement)


### Back port request

### Approach
#### What is the motivation for this PR?
    Make show techsupport UTs run with KVM test.

#### How did you do it?
   Add show_techsupport/test_techsupport.py to tests/kvmtest.sh.

#### How did you verify/test it?
   Make sure all current UT not break during merge validation.

#### Any platform specific information?
   N/A

#### Supported testbed topology if it's a new test case?

### Documentation
